Добавлены javadoc, unit-тесты (junit + mockito), получение выписки по счёту и сохранение её в pdf, получение чеков и сохранение их в txt, получение "money statement"

+++ src/main/java/org/example/model/Transaction.java
@@ -1,6 +1,7 @@
 package org.example.model;
 
 import lombok.Data;
+import lombok.ToString;
 import org.example.enumeration.TransactionType;
 
 import java.math.BigDecimal;
@@ -9,7 +10,7 @@ import java.sql.Time;
 
 @Data
 public class Transaction {
-    private int id;
+    private long id;
     private Account senderAccount;
     private Account recipientAccount;
     private BigDecimal amount;
