+++ src/main/java/org/example/repository/TransactionRepository.java
@@ -4,7 +4,19 @@ import org.example.model.Transaction;
 
 public interface TransactionRepository {
 
+    /**
+     * Creates a transaction in the database.
+     *
+     * @param transaction the transaction object
+     * @return the transaction object with the generated ID.
+     */
     Transaction create(Transaction transaction);
 
-    Transaction findById(int id);
+    /**
+     * Finds a transaction by its ID in the database.
+     *
+     * @param id the transaction ID
+     * @return the transaction object.
+     */
+    Transaction findById(long id);
 }
